Backhunt written as OOP
